Adding the ability to change "Last Updated" field
This commit adds the ability to change the behavior of the "Last Updated"
field at the bottom of the html documents generated by asciidoc.
This will be done by using the "footer-style" new attribute.
This commit adds the ability to change it to the revision date or to skip it.

This commit also adds the documentation of this new feature to the user guide.

Testasciidoc tests are also provided in this commit.